% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{data}
\name{Error}
\alias{Error}
\title{Error class}
\arguments{
\item{name}{Name of the error}

\item{behavior}{Behavior of the error. See Details}
}
\description{
Error class
}
\details{
\strong{Methods}
\describe{
 \item{\code{xx}}
}
}
\section{behavior options}{

\itemize{
 \item stop - stop on error
 \item warning - warning on error
 \item message - message on error
 \item function - use function on error
}
}
\examples{
\dontrun{
(x <- Error$new(name = "hello"))
(y <- Error$new(name = "hello", behavior = function(x) x^2))
(z <- Error$new(name = "hello", message_template = "{{status}} - {{message}}"))
}
}
\keyword{datasets}

