% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-classes.R
\name{error_http}
\alias{error_414}
\alias{error_http}
\title{higher level error wrappers}
\usage{
error_http(x, behavior = "stop", message_template = NULL)

error_414(x, behavior = "stop", message_template = NULL)

error_414(x, behavior = "stop", message_template = NULL)
}
\arguments{
\item{x}{the result of a call via httr, curl, or RCurl}

\item{behavior}{Behavior of the error. See \code{\link{Error}}}

\item{message_template}{A message template, not working yet}
}
\description{
higher level error wrappers
}
\examples{
\dontrun{
res <- GET("https://httpbin.org/status/418")
error_http(res)

res <- GET("https://httpbin.org/status/414")
error_414(res)
}
}

