% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fauxpas-package.R
\docType{package}
\name{fauxpas-package}
\alias{fauxpas-package}
\alias{fauxpas}
\title{HTTP Error Helpers}
\description{
HTTP Error Helpers
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
