% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_error_class.R
\name{find_error_class}
\alias{find_error_class}
\title{Find error classes}
\usage{
find_error_class(status_code)
}
\arguments{
\item{status_code}{(numeric,integer) A status code}
}
\value{
an object of class \code{R6ClassGenerator}. call \code{$new()}
to initialize a new instance
}
\description{
Find error classes
}
\examples{
find_error_class(414)
find_error_class(418)
find_error_class(505)

# initialize the class
find_error_class(418)$new()

# not found
\dontrun{find_error_class(999)}
}
\seealso{
\code{\link[fauxpas]{Error}}, \code{\link[fauxpas]{Error-Classes}}
}
