% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors-children.R
\docType{data}
\name{Error-Classes}
\alias{Error-Classes}
\alias{HTTPAccepted}
\alias{HTTPContinue}
\alias{HTTPCreated}
\alias{HTTPFound}
\alias{HTTPMovedPermanently}
\alias{HTTPMultipleChoices}
\alias{HTTPNoContent}
\alias{HTTPNonAuthoritativeInformation}
\alias{HTTPNotModified}
\alias{HTTPOK}
\alias{HTTPPartialContent}
\alias{HTTPRequestNotFound}
\alias{HTTPRequestURITooLong}
\alias{HTTPResetContent}
\alias{HTTPSeeOther}
\alias{HTTPSwitchProtocol}
\title{Individual error classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HTTPContinue

HTTPSwitchProtocol

HTTPOK

HTTPCreated

HTTPAccepted

HTTPNonAuthoritativeInformation

HTTPNoContent

HTTPResetContent

HTTPPartialContent

HTTPMultipleChoices

HTTPMovedPermanently

HTTPFound

HTTPSeeOther

HTTPNotModified

HTTPRequestURITooLong

HTTPRequestNotFound
}
\description{
These error classes are for each HTTP error, and inherit from
the \code{\link{Error}} class in this package.
}
\details{
In addition to what's available in \code{\link{Error}},
these classes have a single variable \code{mssg} that is the very
verbose complete message describing the HTTP condition in detail.
You can include that message in your condition by using \code{do_verbose}
(see below)

\strong{Methods}
\itemize{
  \item \code{do_verbose(response)}
  \code{response} is any response from \pkg{crul}, \pkg{curl}, or \pkg{httr}
  Execute condition, whether it be message, warning, error, or your
  own custom function.
}
}
\section{Functions}{

\itemize{
 \item \code{\link{HTTPRequestURITooLong}}
 \item \code{\link{HTTPRequestNotFound}}
}
}
\examples{
if (requireNamespace("crul")) {
 library("crul")
 res <- HttpClient$new("https://httpbin.org/status/418")$get()
 x <- HTTPRequestURITooLong$new(behavior = "stop")
 \dontrun{
 x$do(res)
 x$do_verbose(res)
 }
}

}
\keyword{datasets}

