% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{data}
\name{Error}
\alias{Error}
\title{Error class}
\arguments{
\item{behavior}{Behavior of the error. See Details}

\item{call.}{(logical) indicating if the call should become part
of the error message. Default: \code{FALSE}}
}
\description{
Error class
}
\details{
\strong{Methods}
\itemize{
  \item \code{do}
  Execute condition, whether it be message, warning, or error.
}
}
\section{behavior options}{

\itemize{
 \item stop - stop on error
 \item warning - warning on error
 \item message - message on error
}
}
\examples{
# create error classes
(x <- Error$new())
(y <- Error$new(message_template = "{{status}} - {{message}}"))

if (requireNamespace("crul")) {
 library("crul")
 res <- HttpClient$new("https://httpbin.org/status/418")$get()

 # stop
 (x <- Error$new(behavior = "stop"))
 \dontrun{x$do(res)}

 # warn
 (x <- Error$new(behavior = "warning"))
 x$do(res)

 # do vs. do_verbose
 x <- HTTPRequestURITooLong$new(behavior = "stop")
 res <- HttpClient$new("https://httpbin.org/status/414")$get()
 \dontrun{
 error_414(res)
 x$do(res)
 x$do_verbose(res)
 }

 # i'm a teapot
 x <- HTTPTeaPot$new(behavior = "stop")
 res <- HttpClient$new("https://httpbin.org/status/418")$get()
 x$do(res)
 x$do_verbose(res)
}
}
\keyword{datasets}

