% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{data}
\name{Error}
\alias{Error}
\title{Error class}
\arguments{
\item{behavior}{Behavior of the error. See Details}

\item{call.}{(logical) indicating if the call should become part
of the error message. Default: \code{FALSE}}
}
\description{
Error class
}
\details{
\strong{Methods}
\describe{
 \item{\code{xx}}
}
}
\section{behavior options}{

\itemize{
 \item stop - stop on error
 \item warning - warning on error
 \item message - message on error
 \item function - use function on error
}
}
\examples{
\dontrun{
library("httr")

(x <- Error$new())
(y <- Error$new(fun = function(x) x^2))
(z <- Error$new(message_template = "{{status}} - {{message}}"))

res <- GET("https://httpbin.org/status/418")

# stop
(x <- Error$new(behavior = "stop"))
x$do(res)

# warn
(x <- Error$new(behavior = "warning"))
x$do(res)

x <- HTTPRequestURITooLong$new(behavior = "stop")
res <- GET("https://httpbin.org/status/414")
x$do(res)
}
}
\keyword{datasets}

