% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-classes.R
\name{error_http}
\alias{error_101}
\alias{error_404}
\alias{error_414}
\alias{error_418}
\alias{error_http}
\title{higher level error wrappers}
\usage{
error_http(x, behavior = "stop", message_template = NULL)

error_101(x, behavior = "stop", message_template = NULL)

error_404(x, behavior = "stop", message_template = NULL)

error_414(x, behavior = "stop", message_template = NULL)

error_418(x, behavior = "stop", message_template = NULL)
}
\arguments{
\item{x}{the result of a call via \pkg{crul}, \pkg{curl}, or \pkg{httr}}

\item{behavior}{Behavior of the error. See \code{\link{Error}}}

\item{message_template}{A message template, not working yet}
}
\description{
higher level error wrappers
}
\examples{
if (requireNamespace("crul")) {
 library("crul")
 res <- HttpClient$new("https://httpbin.org/status/418")$get()
 \dontrun{error_http(res)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")

 res <- HttpClient$new("https://httpbin.org/status/414")$get()
 \dontrun{error_414(res)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")

 res <- HttpClient$new("https://httpbin.org/asdfafadsf")$get()
 \dontrun{error_404(res)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")
}

if (requireNamespace("curl")) {
 library("curl")
 h <- curl::new_handle()
 curl::handle_setopt(h)
 resp <- curl::curl_fetch_memory("https://httpbin.org/status/418", h)
 \dontrun{error_http(resp)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")
}

if (requireNamespace("httr")) {
 library("httr")
 res <- GET("https://httpbin.org/status/418")
 \dontrun{error_http(res)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")
}
}

