% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_http_generated.R, R/s3-classes.R
\name{error_100}
\alias{error_100}
\alias{error_101}
\alias{error_102}
\alias{error_200}
\alias{error_201}
\alias{error_202}
\alias{error_203}
\alias{error_204}
\alias{error_205}
\alias{error_206}
\alias{error_207}
\alias{error_208}
\alias{error_226}
\alias{error_300}
\alias{error_301}
\alias{error_302}
\alias{error_303}
\alias{error_304}
\alias{error_305}
\alias{error_306}
\alias{error_307}
\alias{error_308}
\alias{error_400}
\alias{error_401}
\alias{error_402}
\alias{error_403}
\alias{error_404}
\alias{error_405}
\alias{error_406}
\alias{error_407}
\alias{error_408}
\alias{error_409}
\alias{error_410}
\alias{error_411}
\alias{error_412}
\alias{error_413}
\alias{error_414}
\alias{error_415}
\alias{error_416}
\alias{error_417}
\alias{error_418}
\alias{error_419}
\alias{error_420}
\alias{error_422}
\alias{error_423}
\alias{error_424}
\alias{error_425}
\alias{error_426}
\alias{error_428}
\alias{error_429}
\alias{error_431}
\alias{error_440}
\alias{error_444}
\alias{error_449}
\alias{error_450}
\alias{error_451}
\alias{error_494}
\alias{error_495}
\alias{error_496}
\alias{error_497}
\alias{error_498}
\alias{error_499}
\alias{error_500}
\alias{error_501}
\alias{error_502}
\alias{error_503}
\alias{error_504}
\alias{error_505}
\alias{error_506}
\alias{error_507}
\alias{error_508}
\alias{error_509}
\alias{error_510}
\alias{error_511}
\alias{error_598}
\alias{error_599}
\alias{error_http}
\title{higher level error wrappers}
\usage{
error_100(x, behavior = "stop", message_template = NULL)

error_101(x, behavior = "stop", message_template = NULL)

error_102(x, behavior = "stop", message_template = NULL)

error_200(x, behavior = "stop", message_template = NULL)

error_201(x, behavior = "stop", message_template = NULL)

error_202(x, behavior = "stop", message_template = NULL)

error_203(x, behavior = "stop", message_template = NULL)

error_204(x, behavior = "stop", message_template = NULL)

error_205(x, behavior = "stop", message_template = NULL)

error_206(x, behavior = "stop", message_template = NULL)

error_207(x, behavior = "stop", message_template = NULL)

error_208(x, behavior = "stop", message_template = NULL)

error_226(x, behavior = "stop", message_template = NULL)

error_300(x, behavior = "stop", message_template = NULL)

error_301(x, behavior = "stop", message_template = NULL)

error_302(x, behavior = "stop", message_template = NULL)

error_303(x, behavior = "stop", message_template = NULL)

error_304(x, behavior = "stop", message_template = NULL)

error_305(x, behavior = "stop", message_template = NULL)

error_306(x, behavior = "stop", message_template = NULL)

error_307(x, behavior = "stop", message_template = NULL)

error_308(x, behavior = "stop", message_template = NULL)

error_400(x, behavior = "stop", message_template = NULL)

error_401(x, behavior = "stop", message_template = NULL)

error_402(x, behavior = "stop", message_template = NULL)

error_403(x, behavior = "stop", message_template = NULL)

error_404(x, behavior = "stop", message_template = NULL)

error_405(x, behavior = "stop", message_template = NULL)

error_406(x, behavior = "stop", message_template = NULL)

error_407(x, behavior = "stop", message_template = NULL)

error_408(x, behavior = "stop", message_template = NULL)

error_409(x, behavior = "stop", message_template = NULL)

error_410(x, behavior = "stop", message_template = NULL)

error_411(x, behavior = "stop", message_template = NULL)

error_412(x, behavior = "stop", message_template = NULL)

error_413(x, behavior = "stop", message_template = NULL)

error_414(x, behavior = "stop", message_template = NULL)

error_415(x, behavior = "stop", message_template = NULL)

error_416(x, behavior = "stop", message_template = NULL)

error_417(x, behavior = "stop", message_template = NULL)

error_418(x, behavior = "stop", message_template = NULL)

error_419(x, behavior = "stop", message_template = NULL)

error_420(x, behavior = "stop", message_template = NULL)

error_422(x, behavior = "stop", message_template = NULL)

error_423(x, behavior = "stop", message_template = NULL)

error_424(x, behavior = "stop", message_template = NULL)

error_425(x, behavior = "stop", message_template = NULL)

error_426(x, behavior = "stop", message_template = NULL)

error_428(x, behavior = "stop", message_template = NULL)

error_429(x, behavior = "stop", message_template = NULL)

error_431(x, behavior = "stop", message_template = NULL)

error_440(x, behavior = "stop", message_template = NULL)

error_444(x, behavior = "stop", message_template = NULL)

error_449(x, behavior = "stop", message_template = NULL)

error_450(x, behavior = "stop", message_template = NULL)

error_451(x, behavior = "stop", message_template = NULL)

error_494(x, behavior = "stop", message_template = NULL)

error_495(x, behavior = "stop", message_template = NULL)

error_496(x, behavior = "stop", message_template = NULL)

error_497(x, behavior = "stop", message_template = NULL)

error_498(x, behavior = "stop", message_template = NULL)

error_499(x, behavior = "stop", message_template = NULL)

error_500(x, behavior = "stop", message_template = NULL)

error_501(x, behavior = "stop", message_template = NULL)

error_502(x, behavior = "stop", message_template = NULL)

error_503(x, behavior = "stop", message_template = NULL)

error_504(x, behavior = "stop", message_template = NULL)

error_505(x, behavior = "stop", message_template = NULL)

error_506(x, behavior = "stop", message_template = NULL)

error_507(x, behavior = "stop", message_template = NULL)

error_508(x, behavior = "stop", message_template = NULL)

error_509(x, behavior = "stop", message_template = NULL)

error_510(x, behavior = "stop", message_template = NULL)

error_511(x, behavior = "stop", message_template = NULL)

error_598(x, behavior = "stop", message_template = NULL)

error_599(x, behavior = "stop", message_template = NULL)

error_http(x, behavior = "stop", message_template = NULL)
}
\arguments{
\item{x}{the result of a call via \pkg{crul}, \pkg{curl}, or \pkg{httr}}

\item{behavior}{Behavior of the error. See \code{\link{Error}}}

\item{message_template}{A message template, not working yet}
}
\description{
higher level error wrappers
}
\examples{
if (requireNamespace("crul")) {
 library("crul")
 res <- HttpClient$new("https://httpbin.org/status/418")$get()
 \dontrun{error_http(res)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")

 res <- HttpClient$new("https://httpbin.org/status/414")$get()
 \dontrun{error_414(res)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")

 res <- HttpClient$new("https://httpbin.org/asdfafadsf")$get()
 \dontrun{error_404(res)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")
}

if (requireNamespace("curl")) {
 library("curl")
 h <- curl::new_handle()
 curl::handle_setopt(h)
 resp <- curl::curl_fetch_memory("https://httpbin.org/status/418", h)
 \dontrun{error_http(resp)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")
}

if (requireNamespace("httr")) {
 library("httr")
 res <- GET("https://httpbin.org/status/418")
 \dontrun{error_http(res)}
 error_http(res, behavior = "warning")
 error_http(res, behavior = "message")
}
}

